melhora na trativa de erro

+++ src/repositories/deleteUser.repositorie.ts
@@ -2,6 +2,16 @@ import { prisma } from "../configs/prisma";
 
 export class DeleteUser {
   async delete(id: string) {
+    const emailExists = await prisma.username.findFirst({
+      where: {
+        id,
+      },
+    });
+
+    if (!emailExists) {
+      return "Id not exists";
+    }
+
     try {
       await prisma.username.delete({
         where: {
